backend/requirements.txt
Django>=5.0.2
djangorestframework>=3.14.0
celery>=5.3.6
redis>=5.0.1
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
gunicorn>=21.2.0
requests>=2.31.0
beautifulsoup4>=4.12.3
openai>=1.12.0
drf-spectacular>=0.27.1
django-celery-results>=2.5.1
